feat: allow only authors to delete its own questions

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -1,6 +1,7 @@
 import { QuestionsRepository } from '../repositories/question-repository'
 
 interface DeleteQuestionUseCaseRequest {
+	authorId: string
 	questionId: string
 }
 
@@ -10,10 +11,12 @@ export class DeleteQuestionUseCase {
     private questionsRepository: QuestionsRepository
 	) { }
 
-	async execute({ questionId }: DeleteQuestionUseCaseRequest): Promise<void> {
+	async execute({ authorId, questionId }: DeleteQuestionUseCaseRequest): Promise<void> {
 		const question = await this.questionsRepository.findById(questionId)
 
 		if(!question) throw new Error('Question not found.')
+
+		if(authorId !== question.authorId.toString()) throw new Error('Not allowed.')
 
 		await this.questionsRepository.delete(question)
 	}
